+++ CHANGELOG.md
@@ -10,6 +10,7 @@ nav_order: 1
 
 - Fix races in tests
 - Add support for normalization of `ip_filter_object` user config options
+- Improve team member deletion
 
 ## [3.10.0] - 2022-12-14
 - Add ClickHouse examples: